fix(migrate): replace ink by prompts select
Closes prisma/migrate#342
