+++ src/packages/migrate/src/utils/test-interactivelyCreateDatabase.ts
@@ -6,8 +6,6 @@ import {
 async function main(): Promise<void> {
   // const result = await interactivelyCreateDatabase('file:dev7.db', 'apply')
   const result = await askToCreateDb('file:dev7.db', 'apply', '.')
-
-  console.log(result)
 }
 
 main()
